requirements.txt
disnake==2.9.1
PyYAML==6.0.1
